Changes to the UI building:
Noticed that the way I thought to make the ui wasn't clever, so I have
to rebuild all the location and game classes. Going to take a while to
fix stuff. Mid-fixing commit, going to take a break. A lot to work on, I
wish I had started doing this in an other way.