Fixing up stuff.
Got the scaleGameApp to draw the sprites. The sprites are in wrong
places but there is logic to the error. also wrong sprites but that is
probably just because I cant count. But a milestone that I finally got
the game to draw stuff. Now I just need to get it to draw the right
stuff and in correct places :p