It works! (somewhat)
Added scoring system, and added visualisation for the player scores.
Just gotta do a turn system and it's a somewhat working product.
Basically all the algorithms work for the scales
and the scoring. Tipping works as well, as well as the chain reaction of
tipping. Final day tomorrow, hopefully am going 
to get the minimum stuff done in time and also the project plan.